Clean Python Packages

+++ app/db.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
-import typing
 import logging
 import threading
+import typing
 
 import pymongo
-import pymongo.database
 import pymongo.collection
+import pymongo.database
 
 from . import commons
 
@@ -33,24 +33,27 @@ class DBManager:
         self.connect()
 
     def disconnect(self):
-        """ Disconnect """
+        """Disconnect"""
         if self.client:
             self.client.close()
 
     def connect(self):
-        """ Estabilish mongo connection """
+        """Estabilish mongo connection"""
         self.client = pymongo.MongoClient(self.url)
         self.mailpy_db: pymongo.database.Database = self.client[DBManager.DB_NAME]
 
     def get_entries(self):
-        """ Return all entries """
+        """Return all entries"""
         entries: pymongo.collection.Collection = self.mailpy_db[
             DBManager.ENTRIES_COLLECTION
         ]
         return [e for e in entries.find()]
 
     def create_entry(self, entry: commons.Entry):
-        """ Create an entry """
+        """Create an entry"""
+        if not self.mailpy_db:
+            raise RuntimeError("Database not initialised")
+
         if not entry.group:
             logger.warning(f"Invalid group for entry {entry}")
             return
@@ -67,32 +70,25 @@ class DBManager:
             DBManager.ENTRIES_COLLECTION
         ]
 
-        entries.create_index(
-            [
-                ("pvname", pymongo.ASCENDING),
-                ("emails", pymongo.ASCENDING),
-                ("condition", pymongo.ASCENDING),
-                ("alarm_values", pymongo.ASCENDING),
-            ],
-            unique=True,
-        )
-
+        logger.info(f"insert {entry.as_dict()}")
         result = entries.insert(entry.as_dict())
         logger.info(f"Inserted entry {entry} id {result}")
 
     def get_group(self, group_name: str) -> typing.Optional[str]:
+        if not self.mailpy_db:
+            raise RuntimeError("Database not initialised")
         groups: pymongo.collection.Collection = self.mailpy_db[
             DBManager.GROUPS_COLLECTION
         ]
         return groups.find_one({"_id": group_name})
 
     def create_group(self, group: commons.Group):
-        """ Create a group """
+        """Create a group"""
+        if not self.mailpy_db:
+            raise RuntimeError("Database not initialised")
         groups: pymongo.collection.Collection = self.mailpy_db[
             DBManager.GROUPS_COLLECTION
         ]
-
-        groups.create_index([("name", pymongo.ASCENDING)], unique=True)
 
         if groups.find_one({"name": group.name}):
             logger.warning(f"Group {group} already exists")
@@ -102,20 +98,25 @@ class DBManager:
         logger.info(f"Inserted group {group} id {result}")
 
     def get_condition(self, name: str):
-        """ Get a condtion by name """
+        """Get a condtion by name"""
+        if not self.mailpy_db:
+            raise RuntimeError("Database not initialised")
+
         conditions: pymongo.collection.Collection = self.mailpy_db[
             DBManager.CONDITIONS_COLLECTION
         ]
         return conditions.find_one({"name": name})
 
     def initialize_conditions(self):
-        """ Initialize the conditions collection using the supported ones from commons.Condition """
+        """Initialize the conditions collection using the supported ones from commons.Condition"""
+        if not self.mailpy_db:
+            raise RuntimeError("Database not initialised")
+
         self.mailpy_db.drop_collection(DBManager.CONDITIONS_COLLECTION)
 
         conditions: pymongo.collection.Collection = self.mailpy_db[
             DBManager.CONDITIONS_COLLECTION
         ]
-        conditions.create_index([("name", pymongo.ASCENDING)], unique=True)
         result = conditions.insert_many(commons.Condition.get_conditions())
 
         logger.info(f"Inserted {result.inserted_ids}")
